chg: Only write file when file not exists.

+++ scripts/goldstandard/file.py
@@ -14,5 +14,8 @@ def write_to_json(url, result_directory, document):
     result_fname = os.path.join(result_directory, f'{url.replace("/", ".")}.json')
     if not os.path.exists(result_directory):
         os.makedirs(result_directory)
-    with open(result_fname, "w") as f2:
-        dump(document, f2, indent=True)
+    if not os.path.isfile(result_fname):
+        with open(result_fname, "w") as f2:
+            dump(document, f2, indent=True)
+    else:
+        logging.warning(f'File {result_fname} already exists')
